Changes to tracing level:
- Changed the default level to be OFF for all subcommands and only for
  the server subcommand to be INFO.
