+++ crates/bld_commands/src/server/command.rs
@@ -3,6 +3,7 @@ use actix_web::rt::System;
 use anyhow::Result;
 use bld_config::BldConfig;
 use clap::Args;
+use tracing::metadata::LevelFilter;
 
 #[derive(Args)]
 #[command(about = "Start bld in server mode, listening to incoming build requests")]
@@ -20,6 +21,14 @@ pub struct ServerCommand {
 impl BldCommand for ServerCommand {
     fn verbose(&self) -> bool {
         self.verbose
+    }
+
+    fn tracing_level(&self) -> LevelFilter {
+        if self.verbose() {
+            LevelFilter::DEBUG
+        } else {
+            LevelFilter::INFO
+        }
     }
 
     fn exec(self) -> Result<()> {
